avoid compiler warnings on 32- and 64-bit systems

+++ parse.c
@@ -957,7 +957,10 @@ static inline void print_event(struct connection *c, const unsigned char *buffer
 		l = 32;
 
 	if( len < l ) {
-		fprintf(stderr, "ignoring incomplete event, length %ld\n", len);
+		fprintf(stderr,
+			"ignoring incomplete event, length %lld\n",
+ 			(long long)len
+		);
 		return;
 	}
